LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/navigating-across-documents/navigate-cross-origin-iframe-to-same-url-expected.txt



FAIL Navigating a cross-origin iframe to its current url should not replace assert_equals: expected 2 but got 3
